HuaweiToDo/requirements.txt
Django==2.2
djangorestframework==3.9.3
numpy==1.16.3
pandas==0.24.2
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
psycopg2-binary==2.8.2
gunicorn==19.9.0